authios::use_cases::user

Struct UsersUseCase

pub struct UsersUseCase;

Implementations

impl UsersUseCase

pub async fn info<'a, A: Acquire<'a, Database = Postgres>>( params: UserInfoParams, client: A, ) -> Result<User, UserInfoError>

UsersUseCase::info

Retrieve a user from JWT token, checking for possible errors

Arguments:

params: crate::params::use_case::user::info::UserInfoParams - the parameters of the query

client: [sqlx::Acquire] - sqlx postgres client

Errors:

Errors described in crate::errors::use_case::user::info::UserInfoError

impl UsersUseCase

pub async fn authorize<'a, A: Acquire<'a, Database = Postgres>>( params: UserAuthorizeParams, client: A, ) -> Result<bool, UserAuthorizeError>

UsersUseCase::authorize

Check if user has patricular permission

Arguments:

params: crate::params::use_case::user::authorize::UserAuthorizeParams - the parameters of the query

client: [sqlx::Acquire] - sqlx postgres client

Errors:

Errors described in crate::errors::use_case::user::authorize::UserAuthorizeError

impl UsersUseCase

pub async fn login<'c, C: Acquire<'c, Database = Postgres>>( params: UserLoginParams, client: C, ) -> Result<String, UserLoginError>

UsersUseCase::login

Log user in and return the session token, checking for possible errors

Arguments:

params: crate::params::use_case::user::login::UserLoginParams - the parameters of the query

client: [sqlx::Acquire] - sqlx postgres client

Errors:

Errors described in crate::errors::use_case::user::login::UserLoginError

impl UsersUseCase

pub async fn update_pwd<'a, A: Acquire<'a, Database = Postgres>>( params: UserUpdatePwdParams, client: A, ) -> Result<(), UserUpdatePwdError>

UsersUseCase::update_pwd

Update user’s password, hashing password and checking for possible errors

Arguments:

params: crate::params::use_case::user::update_pwd::UserUpdatePwdParams - the parameters of the query

client: [sqlx::Acquire] - sqlx postgres client

Errors:

Errors described in crate::errors::use_case::user::update_pwd::UserUpdatePwdError

impl UsersUseCase

pub async fn register<'c, C: Acquire<'c, Database = Postgres>>( params: UserRegisterParams, client: C, ) -> Result<(), UserRegisterError>

UsersUseCase::register

Log user in and return the session token, checking for possible errors

Arguments:

params: crate::params::use_case::user::register::UserRegisterParams - the parameters of the query

client: [sqlx::Acquire] - sqlx postgres client

Errors:

Errors described in crate::errors::use_case::user::register::UserRegisterError

impl UsersUseCase

pub async fn grant_group<'a, A: Acquire<'a, Database = Postgres>>( params: UserGrantGroupParams, client: A, ) -> Result<(), UserGrantGroupError>

UsersUseCase::grant_group

Grant a group to a user, checking for possible errors

Arguments:

params: crate::params::use_case::user::grant_group::UserGrantGroupParams - the parameters of the query

client: [sqlx::Acquire] - sqlx postgres client

Errors:

Errors described in crate::errors::use_case::user::grant_group::UserGrantGroupError

impl UsersUseCase

pub async fn revoke_group<'a, A: Acquire<'a, Database = Postgres>>( params: UserRevokeGroupParams, client: A, ) -> Result<(), UserRevokeGroupError>

UsersUseCase::revoke

Revoke a group from a group, checking for possible errors

Arguments:

params: crate::params::use_case::user::revoke_group::UserRevokeGroupParams - the parameters of the query

client: [sqlx::Acquire] - sqlx postgres client

Errors:

Errors described in crate::errors::use_case::user::revoke_group::UserRevokeGroupError

impl UsersUseCase

pub async fn init_root<'a, A: Acquire<'a, Database = Postgres>>( params: UserInitRootParams, client: A, ) -> Result<(), UserInitRootError>

UsersUseCase::init_root

Init the root account and all of related data checking for errors

Arguments:

params: crate::params::use_case::user::init_root::UserInitRootParams - the parameters of the query

client: [sqlx::Acquire] - sqlx postgres client

Errors:

Errors described in crate::errors::use_case::user::init_root::UserInitRootError

impl UsersUseCase

pub async fn list_permissions<'a, A: Acquire<'a, Database = Postgres>>( params: UserListPermissionsParams, client: A, ) -> Result<Vec<String>, UserListPermissionsError>

UsersUseCase::list_permissions

List all user’s permissions

Arguments:

params: crate::params::use_case::user::list_permissions::UserListPermissionsParams - the parameters of the query

client: [sqlx::Acquire] - sqlx postgres client

Errors:

Errors described in crate::errors::use_case::user::list_permissions::UserListPermissionsError